# apps/api/swagger.yaml
openapi: 3.0.3
info:
  title: HostMyClaw API
  description: |
    REST API for **HostMyClaw** — a SaaS platform for hosting OpenClaw AI assistants.

    ## Authentication
    Most endpoints require a Bearer JWT token. Obtain one via `POST /api/v1/auth/login`
    or `POST /api/v1/auth/register`.

    Include in every protected request:
    ```
    Authorization: Bearer <accessToken>
    ```

    Tokens expire in **15 minutes**.

    ## Plans
    | Plan     | Assistants | Price/month |
    |----------|-----------|-------------|
    | free     | 0         | $0          |
    | starter  | 1         | $29         |
    | growth   | 5         | $79         |
    | pro      | 999       | $199        |

    ## Assistant lifecycle
    `draft` → `queued` → `deploying` → `running` → `stopping` → `stopped`
    Any state can transition to `failed`.
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Local development
  - url: https://api.yourdomain.com
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT obtained from /api/v1/auth/login or /api/v1/auth/register
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Not found
      required: [error]
    AppError:
      type: object
      properties:
        code:
          type: string
          example: LIMIT_EXCEEDED
        error:
          type: string
          example: Assistant limit reached for your plan
      required: [code, error]

    # ── Domain models ────────────────────────────────────────────────────────────
...
    Assistant:
      type: object
      properties:
        id:
          type: string
          example: asst_7pL3kR9tY
        user_id:
          type: string
          example: usr_4xK9mB2nQ
        name:
          type: string
          example: My Sales Bot
        template_id:
          type: string
          nullable: true
          example: tmpl_sales
        status:
          type: string
          enum: [draft, queued, deploying, running, stopping, stopped, failed]
          example: running
        config:
          type: object
          additionalProperties: true
          example: { language: "en", maxTokens: 2048 }
        container_name:
          type: string
          nullable: true
          example: openclaw_asst7pl3kr9ty
        container_id:
          type: string
          nullable: true
        container_port:
          type: integer
          nullable: true
          example: 8501
        subdomain:
          type: string
          nullable: true
          example: asst7pl3kr9ty
        health_status:
          type: string
          nullable: true
          example: healthy
        last_health_at:
          type: string
          format: date-time
          nullable: true
        uptime_since:
          type: string
          format: date-time
          nullable: true
        error_message:
          type: string
          nullable: true
        deployed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, user_id, name, status, created_at, updated_at]
...
paths:
...
  /api/v1/assistants/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        example: asst_7pL3kR9tY

    get:
      tags: [Assistants]
      summary: Get assistant by ID
      operationId: getAssistant
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Assistant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assistant"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      tags: [Assistants]
      summary: Update assistant name or config
      operationId: updateAssistant
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  example: Renamed Bot
                config:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Updated assistant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assistant"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags: [Assistants]
      summary: Delete assistant
      operationId: deleteAssistant
      description: If the assistant has a running container, a teardown job is enqueued before deletion.
      security:
        - BearerAuth: []
      responses:
        "204":
          description: Deleted — no content
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/assistants/{id}/deploy:
    post:
      tags: [Assistants]
      summary: Deploy assistant to a container
      operationId: deployAssistant
      description: |
        Enqueues a deployment job. The assistant must be in `draft`, `stopped`, or `failed` status.

        **Requires** an active paid subscription with at least one free assistant slot.
        On success the assistant transitions to `queued` and a `jobId` is returned.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: asst_7pL3kR9tY
      responses:
        "202":
          description: Deployment enqueued
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId:
                    type: string
                    example: job_xM7nK2pQ
                  assistantId:
                    type: string
                    example: asst_7pL3kR9tY
                  status:
                    type: string
                    example: queued
        "400":
          description: Assistant is in an invalid state for deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
        "403":
          description: Subscription required or assistant limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
        "404":
          description: Assistant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: No container ports available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
  /api/v1/assistants/{id}/stop:
    post:
      tags: [Assistants]
      summary: Stop a running assistant
      operationId: stopAssistant
      description: The assistant must be in `running` status. It transitions to `stopping`.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stop initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: stopping
        "400":
          description: Assistant is not running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/assistants/{id}/restart:
    post:
      tags: [Assistants]
      summary: Restart an assistant container
      operationId: restartAssistant
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Restart initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: restarting
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...